requirements.txt
slackclient==1.3.0
python-ldap
pyyaml